Fixed bug in SymbolicVariable by reverting it back to r296.
In SymbolicVariable propagation was even successfull when no solution
was left and all internal variables where set to false.
In the changes to revision 297 we probably made a mistake when we
wanted to say that exactly one has to be true but only said that not
more than one can be true.
From the changes after r296 I only left in the call to
addConstraintNoPropagation.